+++ src/frontend/Orion/Configuration/AdminSettings/NodeManagement/NodeManagement.js
@@ -1,0 +1,23 @@
+import React from 'react'
+
+import NodeMgmtHeader from './NodeMgmtHeader'
+import ToolsPanel from './ToolsPanel'
+import DashboardsPanel from './DashboardsPanel'
+import PagesPanel from './PagesPanel'
+import CardsPanel from './CardsPanel'
+
+const NodeManagement = () => {
+  return (
+    <div style={{ flex: 1 }}>
+      <NodeMgmtHeader />
+      <div style={{ display: 'flex' }}>
+        <ToolsPanel />
+        <DashboardsPanel />
+        <PagesPanel />
+        <CardsPanel />
+      </div>
+    </div>
+  )
+}
+
+export default NodeManagement